trilha service completada

+++ src/main/java/br/com/devschool/devschool/repository/AlunoRepository.java
@@ -1,6 +1,7 @@
 package br.com.devschool.devschool.repository;
 
 import br.com.devschool.devschool.model.Aluno;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlunoRepository extends JpaRepository<Aluno,Integer> {
+public interface AlunoRepository extends JpaRepository<Aluno, Long> {
 }
